requirements.txt
Flask
bs4
python-dotenv
pyzbar
opencv-python
requests
lxml